feat: moved footer & navbar to /shared/components
