+++ src/app/shared/components/footer/footer.component.ts
